Move back to previous scrolling position after prepending older messages.
Extract a Messages component to make this much easier. This also further cleans up the chat route.
This solution is still a bit ugly as it is confusing to see the container briefly scroll to the top and then back again. It remains to be seen how we can prevent the brief scroll to the top.