feature/EVG-222: Fix scrolling behavior in chat (#77)
* extract scrolling functions, scroll to bottom upon groupUuid change

* update svelte

* increase padding on join call loading spinner to prevent jumping surroundings

* Move back to previous scrolling position after prepending older messages.

Extract a Messages component to make this much easier. This also further cleans up the chat route.
This solution is still a bit ugly as it is confusing to see the container briefly scroll to the top and then back again. It remains to be seen how we can prevent the brief scroll to the top.

* remove debugging print statements